+++ app/static/js/layers.js
@@ -1,3 +1,6 @@
+const website_url = "http://se4allwebpage.westeurope.cloudapp.azure.com";
+
+
 // Basic png-tile layer taken from OpenStreetMap
 var osm = L.tileLayer("https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png", {
   maxZoom: 19,
@@ -20,25 +23,25 @@ var esri = L.tileLayer("https://server.arcgisonline.com/ArcGIS/rest/services/Wor
 // Basic png-tile taken from wmflabs
 var osm_gray = L.tileLayer("https://tiles.wmflabs.org/bw-mapnik/{z}/{x}/{y}.png", {
   maxZoom: 19,
-  attribution: '&copy; Tiles style from OpenStreetMap, hosted by wmflabs. <a href="https://wmflabs.org"></a>'
+  attribution: '&copy OpenStreetMap, <a href="https://wmflabs.org">wmflabs</a>.'
 });
 
 // Basic png-tile layer for background taken from tile server serves zoom levels 5-9
 var national_background = L.tileLayer(tileserver + "nesp2_national_background/{z}/{x}/{y}.png", {
   maxZoom: 19,
-  attribution: '☮'
+  attribution: 'Background Tiles <a href="' + website_url +  '/about-map"> &copy se4all</a>'
 });
 
 // Basic png-tile layer for overlays taken from tile server serves zoom levels 5-9
 var national_heatmap = L.tileLayer(tileserver + "nesp2_national_heatmap/{z}/{x}/{y}.png", {
   maxZoom: 19,
-  attribution: ''
+  attribution: 'Heatmap <a href="' + website_url +  '/about-map">© SE4ALL</a>'
 });
 
 // Basic png-tile layer for overlays taken from tile server serves zoom levels 5-9
 var national_grid = L.tileLayer(tileserver + "nesp2_national_grid/{z}/{x}/{y}.png", {
   maxZoom: 19,
-  attribution: ''
+  attribution: 'Grid <a href="' + website_url +  '/about-map">© SE4ALL</a>'
 });
 
 // Basic png-tile layer combines national grid, heatmap and background. Redundant. Serves Levels 5-9
@@ -50,15 +53,27 @@ var welcome_view = L.tileLayer(tileserver + "nesp2_national_welcome-view/{z}/{x}
 var centroidsGroup = L.layerGroup().addTo(map);
 
 
-function remove_layer(layer) {
+function remove_layer(layer, msg=null) {
   if (map.hasLayer(layer) == true) {
     map.removeLayer(layer);
   }
+  else{
+    if (msg != null){
+        console.log("Cannot remove unexisting layer");
+        console.log(msg);
+    }
+  }
 };
 
-function add_layer(layer) {
+function add_layer(layer, msg=null) {
   if (map.hasLayer(layer) == false) {
     map.addLayer(layer);
+  }
+  else{
+    if (msg != null){
+        console.log("Cannot add already existing layer");
+        console.log(msg);
+    }
   }
 };
 
@@ -104,8 +119,6 @@ function highlight_state(feature, layer) {
     // Update the name of the selected state only if different from the currently selected
     if (selectedState != feature.properties["name"]) {
       prevState = selectedState;
-      //update availability of state for display in side menu
-      selectedStateAvailability = feature.properties.availability;
       //update the selected state
       selectedState = feature.properties["name"];
       // Update the dropdown menu for state selection
@@ -168,16 +181,18 @@ function update_nigeria_states_borders_geojson() {
 };
 
 function zoomToSelectedState() {
-    map.flyToBounds(selectedStateOptions.bounds, {maxZoom: 19});
+    map.flyToBounds(selectedStateOptions.bounds, {maxZoom: 19, animate:false});
 };
 
-// Definitions and functions for the grid_layer
+// Definitions and functions for the state_grid_layer
 
 // Vector tiles layer that is adapted (URL) depending on the chosen state. Contains layers '11kV' and '33kV' Columns: several, but of no interest to the map
 
 //Always Using the entire Grid
-var grid_layer = L.vectorGrid.protobuf(tileserver + "nesp2_state_grid/{z}/{x}/{y}.pbf", {
+var state_grid_layer = L.vectorGrid.protobuf(tileserver + "nesp2_state_grid/{z}/{x}/{y}.pbf", {
   rendererFactory: L.canvas.tile,
+  maxZoom:15,
+  attribution: 'Grid <a href="' + website_url +  '/about-map">© SE4ALL</a>',
   vectorTileLayerStyles: {
     '33_kV': function(prop, zoom) {
       return gridStyle33kv
@@ -187,32 +202,6 @@ var grid_layer = L.vectorGrid.protobuf(tileserver + "nesp2_state_grid/{z}/{x}/{y
     },
   }
 });
-
-// Assign the selected state grid tile to the grid_layer
-function redefine_grid_layer() {
-// Vector tiles layer that is adapted (URL) depending on the chosen state. Contains layers '11kV' and '33kV' Columns: several, but of no interest to the map
-
-//Always Using the entire Grid
-  grid_layer = L.vectorGrid.protobuf(tileserver + "nesp2_state_grid/{z}/{x}/{y}.pbf", {
-    rendererFactory: L.canvas.tile,
-    vectorTileLayerStyles: {
-      '33_kV': function(prop, zoom) {
-        return gridStyle33kv
-      },
-      '11_kV': function(prop, zoom) {
-        return gridStyle11kv
-      },
-    }
-  });
-};
-
-// Update the state level grid layer with tiles
-function update_grid_layer() {
-  //remove_layer(grid_layer);
-  //redefine_grid_layer();
-  // Add the grid layer depending on grid checkbox value
-  grid_cb_fun();
-};
 
 // Adds functions for filters and styling to a defined input grid-cluster-Layer
 function addFunctionsToClusterLayer(layer) {
@@ -248,7 +237,7 @@ function addFunctionsToClusterLayer(layer) {
     map.flyToBounds([
       [e.layer.properties.bb_south, e.layer.properties.bb_west],
       [e.layer.properties.bb_north, e.layer.properties.bb_east]
-    ])
+    ],{animate:false})
   });
   layer.filter = function(filter) {
     let newhiddenIDs = [];
@@ -338,6 +327,7 @@ function createNewClusterLayer(clusterString) {
     minZoom: 5,
     interactive: true,
     tolerance: 10,
+    attribution: 'Settlements <a href="' + website_url +  '/about-map">© SE4ALL</a>',
     getFeatureId: function(f) {
       if (f.properties.cluster_all_id !== undefined) {
         return f.properties.cluster_all_id;
@@ -395,7 +385,7 @@ function addFunctionsToOGClusterLayer(layer) {
     map.flyToBounds([
       [e.layer.properties.bb_south, e.layer.properties.bb_west],
       [e.layer.properties.bb_north, e.layer.properties.bb_east]
-    ])
+    ],{animate:false})
   });
   layer.filter = function(filter) {
     let newhiddenIDs = [];
@@ -477,6 +467,7 @@ function createNewOGClusterLayer(ogClusterString) {
     minZoom: 5,
     interactive: true,
     tolerance: 10,
+    attribution: 'Settlements <a href="' + website_url +  '/about-map">© SE4ALL</a>',
     getFeatureId: function(f) {
       if (f.properties.cluster_offgrid_id !== undefined) {
         return f.properties.cluster_offgrid_id;